backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.9",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "playwright>=1.52.0",
    "requests>=2.32.3",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",  # For handling form data
    "openai>=1.14.0",  # For OpenRouter API compatibility
    "python-dotenv>=1.0.1",  # For environment variables
    "pillow>=11.1.0",  # For image processing
]
